+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Laravel-Artisan Cookbook Changelog
 
+## 0.3.0
+* Added support of `php artisan cache:clear`
+* Added support of `php artisan config:cache`
+* Added support of `php artisan config:clear`
+* Added support of `php artisan clear-compiled`
+* Added support of `php artisan env`
+* Added support of `php artisan package:discover`
+* Extracted each `php artisan` command in a variable.
+
+
 ## 0.2.0
 * Checking existance of the artisan file.
 * Added documentation and other required files.